requirements-1.txt
python-telegram-bot==21.0
httpx==0.25.0
